+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
-- Attributes#genderize
+- Attributes#otherize
+- Allow gamefic-4.0
 
 ## 1.0.0 - September 30, 2024
 - First release